simplify architectures
