+++ dacapo/experiments/architectures/__init__.py
@@ -1,7 +1,5 @@
-from .architecture import Architecture  # noqa
-from .architecture_config import ArchitectureConfig  # noqa
+from .architecture import ArchitectureConfig  # noqa
 from .dummy_architecture_config import (
     DummyArchitectureConfig,
-    DummyArchitecture,
 )  # noqa
-from .cnnectome_unet_config import CNNectomeUNetConfig, CNNectomeUNet  # noqa
+from .cnnectome_unet import CNNectomeUNetConfig  # noqa
