+++ dacapo/experiments/tasks/predictors/hot_distance_predictor.py
@@ -123,11 +123,11 @@ class HotDistancePredictor(Predictor):
         Examples:
             >>> model = predictor.create_model(architecture)
         """
-        if architecture.dims == 2:
+        if architecture.input_shape.dims == 2:
             head = torch.nn.Conv2d(
                 architecture.num_out_channels, self.embedding_dims, self.kernel_size
             )
-        elif architecture.dims == 3:
+        elif architecture.input_shape.dims == 3:
             head = torch.nn.Conv3d(
                 architecture.num_out_channels, self.embedding_dims, self.kernel_size
             )
